requirements.txt
attrdict==2.0.0
gevent==1.4.0
gevent-websocket==0.10.1
h5py==2.6.0
python-engineio==3.11.2
python-socketio==4.4.0
numpy==1.13.1
Pillow==6.2.1
scipy==0.19.1
keras==2.0.8
tensorflow==1.3.0
ujson==1.35
wsaccel==0.6.2


